desktop_app/requirements.txt
openpyxl
tzdata
pillow
